+++ src/transformers/models/beit/modeling_beit.py
@@ -33,7 +33,8 @@ from ...modeling_outputs import (
     SemanticSegmenterOutput,
     SequenceClassifierOutput,
 )
-from ...modeling_utils import PreTrainedModel, find_pruneable_heads_and_indices, prune_linear_layer
+from ...modeling_utils import PreTrainedModel, prune_linear_layer
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import (
     add_code_sample_docstrings,
     add_start_docstrings,
