+++ src/transformers/models/decision_transformer/modeling_decision_transformer.py
@@ -25,7 +25,8 @@ from packaging import version
 from torch import nn
 
 from ...activations import ACT2FN
-from ...modeling_utils import Conv1D, PreTrainedModel, find_pruneable_heads_and_indices, prune_conv1d_layer
+from ...modeling_utils import Conv1D, PreTrainedModel, prune_conv1d_layer
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import (
     ModelOutput,
     add_start_docstrings,
