+++ src/transformers/models/distilbert/modeling_distilbert.py
@@ -42,9 +42,9 @@ from ...modeling_outputs import (
 from ...modeling_utils import (
     PreTrainedModel,
     apply_chunking_to_forward,
-    find_pruneable_heads_and_indices,
     prune_linear_layer,
 )
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import (
     add_code_sample_docstrings,
     add_start_docstrings,
