+++ src/transformers/models/dpt/modeling_dpt.py
@@ -42,7 +42,8 @@ from ...modeling_outputs import (
     DepthEstimatorOutput,
     SemanticSegmenterOutput,
 )
-from ...modeling_utils import PreTrainedModel, find_pruneable_heads_and_indices, prune_linear_layer
+from ...modeling_utils import PreTrainedModel, prune_linear_layer
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import logging
 from .configuration_dpt import DPTConfig
