+++ src/transformers/models/gpt2/modeling_gpt2.py
@@ -44,9 +44,9 @@ from ...modeling_utils import (
     Conv1D,
     PreTrainedModel,
     SequenceSummary,
-    find_pruneable_heads_and_indices,
     prune_conv1d_layer,
 )
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import (
     ModelOutput,
     add_code_sample_docstrings,
