+++ src/transformers/models/longformer/modeling_longformer.py
@@ -27,9 +27,9 @@ from ...activations import ACT2FN, gelu
 from ...modeling_utils import (
     PreTrainedModel,
     apply_chunking_to_forward,
-    find_pruneable_heads_and_indices,
     prune_linear_layer,
 )
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import (
     ModelOutput,
     add_code_sample_docstrings,
