+++ src/transformers/models/mobilebert/modeling_mobilebert.py
@@ -41,7 +41,8 @@ from ...modeling_outputs import (
     SequenceClassifierOutput,
     TokenClassifierOutput,
 )
-from ...modeling_utils import PreTrainedModel, find_pruneable_heads_and_indices, prune_linear_layer
+from ...modeling_utils import PreTrainedModel, prune_linear_layer
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import (
     ModelOutput,
     add_code_sample_docstrings,
