+++ src/transformers/models/openai/modeling_openai.py
@@ -32,9 +32,9 @@ from ...modeling_utils import (
     Conv1D,
     PreTrainedModel,
     SequenceSummary,
-    find_pruneable_heads_and_indices,
     prune_conv1d_layer,
 )
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import (
     ModelOutput,
     add_code_sample_docstrings,
