+++ src/transformers/models/perceiver/modeling_perceiver.py
@@ -32,9 +32,9 @@ from ...modeling_outputs import BaseModelOutputWithCrossAttentions
 from ...modeling_utils import (
     PreTrainedModel,
     apply_chunking_to_forward,
-    find_pruneable_heads_and_indices,
     prune_linear_layer,
 )
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import (
     ModelOutput,
     add_start_docstrings,
