+++ src/transformers/models/splinter/modeling_splinter.py
@@ -28,9 +28,9 @@ from ...modeling_outputs import BaseModelOutputWithPastAndCrossAttentions, Quest
 from ...modeling_utils import (
     PreTrainedModel,
     apply_chunking_to_forward,
-    find_pruneable_heads_and_indices,
     prune_linear_layer,
 )
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import add_code_sample_docstrings, add_start_docstrings, add_start_docstrings_to_model_forward, logging
 from .configuration_splinter import SplinterConfig
