+++ src/transformers/models/t5/modeling_t5.py
@@ -33,7 +33,8 @@ from ...modeling_outputs import (
     Seq2SeqLMOutput,
     Seq2SeqModelOutput,
 )
-from ...modeling_utils import PreTrainedModel, find_pruneable_heads_and_indices, prune_linear_layer
+from ...modeling_utils import PreTrainedModel, prune_linear_layer
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import (
     DUMMY_INPUTS,
     DUMMY_MASK,
