+++ src/transformers/models/tapas/modeling_tapas.py
@@ -31,9 +31,9 @@ from ...modeling_outputs import BaseModelOutput, BaseModelOutputWithPooling, Mas
 from ...modeling_utils import (
     PreTrainedModel,
     apply_chunking_to_forward,
-    find_pruneable_heads_and_indices,
     prune_linear_layer,
 )
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import (
     ModelOutput,
     add_start_docstrings,
