+++ src/transformers/models/vit/modeling_vit.py
@@ -26,7 +26,8 @@ from torch.nn import BCEWithLogitsLoss, CrossEntropyLoss, MSELoss
 
 from ...activations import ACT2FN
 from ...modeling_outputs import BaseModelOutput, BaseModelOutputWithPooling, MaskedLMOutput, SequenceClassifierOutput
-from ...modeling_utils import PreTrainedModel, find_pruneable_heads_and_indices, prune_linear_layer
+from ...modeling_utils import PreTrainedModel, prune_linear_layer
+from ...pytorch_utils import find_pruneable_heads_and_indices
 from ...utils import (
     add_code_sample_docstrings,
     add_start_docstrings,
